PRO-3467: Fix README's stale with_tool (singular) references; use a real JSON-mappable type in the schema example
- `RubyLLM::Chat#with_tool` (singular) doesn't exist in 2.0 -- only
  `with_tools(*tools)`, which accepts one or many. Caught by actually
  running the README's own examples against the installed 2.0.0.rc4 gem
  (RubyLLM.configure + a real chat) rather than trusting the earlier
  RSpec-only verification, which never exercised chat.with_tool directly.
- Softened the "is itself released as a prerelease" line to "ships as a
  matching prerelease itself" -- it hasn't actually been released yet
  (matches the CHANGELOG's [Unreleased] fix from the last commit).
- Swapped the Axn-class schema example's `exposes :confidence, type:
  Numeric` for `type: Float`: axn has no JSON Schema mapping for the
  abstract `Numeric` class, so it silently produced an unconstrained `{}`
  property. `Float` maps to `{type: "number"}` as the example implies.
  Same fix in ask_spec.rb's matching test.

+++ spec/axn/ruby_llm/ask_spec.rb
@@ -136,7 +136,7 @@ RSpec.describe Axn::RubyLLM::Ask do
           include Axn
 
           exposes :company_id, type: Integer
-          exposes :confidence, type: Numeric
+          exposes :confidence, type: Float
           def call; end
         end
       end
